extracted java elements resolution in own class
